requirements.txt
asgiref==3.8.1
click==8.1.7
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.3
django-filter==24.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
filelock==3.13.3
gunicorn==21.2.0
h11==0.14.0
packaging==24.0
pillow==10.3.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
uvicorn==0.29.0
virtualenv==20.25.1
